Big Refactor to headless operation
- Only use YAML for config
- Remove get_status methods and get_config_tenmplate
- Refactor Autoscaler into it's own class and use instance of it.
