+++ Procfile
@@ -1,0 +1,1 @@
+bundle exec bin/dynosaur config.yaml
