+++ dynosaur.gemspec
@@ -25,6 +25,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "rspec"
   spec.add_development_dependency "rspec-mocks"
   spec.add_development_dependency "timecop"
+  spec.add_development_dependency "faker"
 
   spec.add_dependency 'activesupport'
   spec.add_dependency "heroku-api" # We should remove that at some point and fully use platform API
