+++ lib/dynosaur/controllers/abstract_controller_plugin.rb
@@ -1,4 +1,3 @@
-require 'librato/metrics'
 
 module Dynosaur
   module Controllers
@@ -15,10 +14,7 @@ module Dynosaur
         @current = nil
         @heroku_app_name = config['heroku_app_name']
         @heroku_api_key = config['heroku_api_key']
-        @librato_email = config['librato_email']
-        @librato_api_key = config['librato_api_key']
         @dry_run = config.fetch("dry_run", false)
-        @stats_callback = self.method(:librato_send) # default built-in stats callback
         @last_results = {}
         @last_change_ts = nil
 
@@ -33,9 +29,9 @@ module Dynosaur
       end
 
       def load_input_plugin(input_plugin_config)
-        # Load the class and instanciate it
+        # Load the class and instantiate it
         begin
-          klass = Kernel.const_get(input_plugin_config['type'])
+          klass = input_plugin_config['type'].constantize
           puts "Instantiating #{klass.name} for config '#{input_plugin_config["name"]}'"
           return klass.new(input_plugin_config)
         rescue NameError => e
@@ -46,18 +42,12 @@ module Dynosaur
       def heroku_manager
         raise NotImplementedError.new("You must define heroku_manager in your controller")
       end
-
       def get_combined_estimate
         estimates = []
-        details = {}
         # Get the estimated dynos from all configured plugins
         @input_plugins.each { |plugin|
-          plugin_status = plugin.get_status
-          details[plugin.name] = plugin_status
-          estimates << plugin_status['estimate']
+          estimates << plugin.estimated_resources
         }
-        @last_results = details
-
 
         # Combine the estimates and mo
         combined_estimate = estimates.max
@@ -76,90 +66,30 @@ module Dynosaur
         heroku_manager.get_current_value
       end
 
-      def get_status
-        status = {
-          "time" => Time.now,
-          "name" => @name,
-          "current" => @current,
-          "current_estimate" => @current_estimate,
-          "last_changed" => @last_change_ts,
-          "results" => @last_results,
-        }
-        return status
+      def get_stats
       end
 
       def run
-        now = Time.now
-
         before = get_current_resource
         @current_estimate = get_combined_estimate
 
         if @current_estimate != before
           scale
         end
-        after = get_current_resource
 
-        if before != after
-          puts "CHANGE: #{before} => #{after}"
+        if before != @current_estimate
+          puts "CHANGE: #{before} => #{@current_estimate}"
           @last_change_ts = Time.now
         end
-        @current = after
-        log_status(now)
-
-        handle_stats(now, @current_estimate, before, after)
-      end
-
-      def log_status(now)
-        details = ""
-        @last_results.each { |name, result|
-          details += "#{name}: #{result["value"]}, #{result["estimate"]}; "
-        }
-        puts "#{now} [combined: #{@current_estimate}]  #{details}"
-      end
-
-      # Built-in stats callback: librato
-      def librato_send(stats)
-        if @librato_api_key.nil? || @librato_api_key.empty? || @librato_email.nil? || @librato_email.empty?
-          puts "No librato api key and email"
-          return
-        end
-        begin
-          Librato::Metrics.authenticate(@librato_email, @librato_api_key)
-
-          metrics = {}
-          stats[:plugins].keys.sort.each { |name|
-            result = stats[:plugins][name]
-            metrics["dynosaur.#{@heroku_app_name}.#{name}.value"] = result["value"]
-            metrics["dynosaur.#{@heroku_app_name}.#{name}.estimate"] = result["estimate"]
-          }
-          metrics["dynosaur.#{@heroku_app_name}.combined.actual"] = stats[:after]
-          metrics["dynosaur.#{@heroku_app_name}.combined.estimate"] = stats[:estimate]
-
-          Librato::Metrics.submit(metrics)
-        rescue Exception => e
-          puts "Error sending librato metrics"
-          puts e.message
-        end
-      end
-
-
-      def handle_stats(now, combined_estimate, before, after)
-        stats = {
-          :plugins => @last_results, # try to minimize race conditions in the iteration
-          :ts => now,
-          :estimate => combined_estimate,
-          :before => before,
-          :after => after
-        }
-        if !@stats_callback.nil?
-          @stats_callback.call(stats)
-        end
+        @current = @current_estimate
       end
 
       #
       # We don't have access to #blank? and it would be dumb to include activesupport
       # for only one method
       #
+      # FIXME: just use config.fetch(name, default) instead of this
+      # (make sure we're not setting empty strings!)
       def default_value_if_blank(value, default)
         # Stolen from: http://api.rubyonrails.org/classes/Object.html#method-i-blank-3F
         empty = value.respond_to?(:empty?) ? !!value.empty? : !value
