+++ lib/dynosaur/inputs/random_plugin.rb
@@ -27,12 +27,6 @@ module Dynosaur
         return (value / 2.0).ceil
       end
 
-      def self.get_config_template
-        {
-          "seed" => ["text"]
-        }
-      end
-
     end
   end
 end
