+++ lib/dynosaur/inputs/rediscloud_memory_usage_input_plugin.rb
@@ -23,14 +23,6 @@ module Dynosaur
         return suitable_plans(value).first
       end
 
-      def self.get_config_template
-        {
-          "max_percentage_threshold" => ["text"],
-          "component_id" => ["text"],
-          "new_relic_api_key" => ["text"],
-        }
-      end
-
       private
 
       def get_memory_usage
