+++ spec/lib/config_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 describe "Loading plugins" do
   it "Should load plugin classes" do
     config = get_config_with_test_plugin(1)
-    Dynosaur.initialize(config)
+    scaler = Dynosaur::Autoscaler.new(config)
     subclasses = Dynosaur::Controllers::AbstractControllerPlugin.subclasses
     subclasses.should include(Dynosaur::Controllers::DynosControllerPlugin)
 
@@ -13,14 +13,14 @@ describe "Loading plugins" do
 
   it "should handle global config" do
     config = get_config_with_test_plugin
-    Dynosaur.initialize(config)
-    Dynosaur.heroku_app_name.should eql config["scaler"]["heroku_app_name"]
+    scaler = Dynosaur::Autoscaler.new(config)
+    scaler.heroku_app_name.should eql config["scaler"]["heroku_app_name"]
   end
 
   it "should configure the random plugin" do
     config = get_config_with_test_plugin
-    Dynosaur.initialize(config)
-    controller_plugins = Dynosaur.controller_plugins
+    scaler = Dynosaur::Autoscaler.new(config)
+    controller_plugins = scaler.controller_plugins
     controller_plugins.length.should eql 1
     controller_plugins[0].input_plugins[0].unit.should eql "randoms"
     controller_plugins[0].input_plugins[0].seed.should eql config["controller_plugins"][0]["input_plugins"][0]["seed"]
@@ -30,7 +30,7 @@ describe "Loading plugins" do
     config = get_config_with_test_plugin
     config["controller_plugins"][0].delete("name")
     expect {
-      Dynosaur.initialize(config)
+      scaler = Dynosaur::Autoscaler.new(config)
     }.to raise_error("You must specify a name")
   end
 end
