+++ spec/lib/dynosaur/autoscaler_spec.rb
@@ -3,7 +3,7 @@
 
 require 'spec_helper'
 
-describe "scaler" do
+describe Dynosaur::Autoscaler do
 
   context "given multiple plugin configs" do
     before do
@@ -11,12 +11,12 @@ describe "scaler" do
 
       @config["controller_plugins"][0]['input_plugins'][0]["interval"] = 0.1
       @config["controller_plugins"][0]['input_plugins'][1]["interval"] = 0.1
-      Dynosaur.initialize(@config)
+      @scaler = Dynosaur::Autoscaler.new(@config)
     end
 
 
     it "should pick the maximum estimate" do
-      dyno_controller = Dynosaur.controller_plugins[0]
+      dyno_controller = @scaler.controller_plugins[0]
       dyno_controller.input_plugins[0].stub(:retrieve) { 10 }
       dyno_controller.input_plugins[1].stub(:retrieve) { 33 }
       combined, details = dyno_controller.get_combined_estimate
@@ -26,7 +26,7 @@ describe "scaler" do
 
     it "should obey max_resource" do
       # Check that we are constrained by max_resource
-      dyno_controller = Dynosaur.controller_plugins[0]
+      dyno_controller = @scaler.controller_plugins[0]
       dyno_controller.input_plugins[0].stub(:retrieve) { @config["controller_plugins"][0]["max_resource"]*4 }
       dyno_controller.input_plugins[1].stub(:retrieve) { 27 }
       combined, details = dyno_controller.get_combined_estimate
@@ -38,12 +38,21 @@ describe "scaler" do
     it "should obey min_resource" do
       # Check that we are constrained by min_resource
 
-      dyno_controller = Dynosaur.controller_plugins[0]
+      dyno_controller = @scaler.controller_plugins[0]
       dyno_controller.input_plugins[0].stub(:retrieve) { 0 }
       dyno_controller.input_plugins[1].stub(:retrieve) { 1 }
       combined, details = dyno_controller.get_combined_estimate
       puts "#{combined} - #{details}"
       combined.should eql @config["controller_plugins"][0]["min_resource"]
+    end
+
+    it 'should send stats' do
+      handler = @scaler.instance_variable_get("@stats_handlers").first
+      expect(handler).to receive(:report).with(@config["scaler"]["heroku_app_name"],
+                                               [instance_of(Dynosaur::Inputs::RandomPlugin),
+                                                instance_of(Dynosaur::Inputs::RandomPlugin)],
+                                               instance_of(Fixnum), instance_of(Fixnum)).and_call_original
+      @scaler.run_loop
     end
   end
 
@@ -52,15 +61,15 @@ describe "scaler" do
       @config = get_config_with_test_plugin(1)
 
       @config["controller_plugins"][0]["interval"] = 0.1
-      Dynosaur.initialize(@config)
+      @scaler = Dynosaur::Autoscaler.new(@config)
     end
 
     it "should report the error" do
-      dyno_controller = Dynosaur.controller_plugins[0]
+      dyno_controller = @scaler.controller_plugins[0]
       rand = dyno_controller.input_plugins[0]
       expect(rand).to receive(:retrieve).at_least(:once).and_raise(StandardError.new "Oh Noes!")
       expect(Dynosaur::ErrorHandler).to receive(:handle).at_least(:once)
-      dyno_controller.get_combined_estimate
+      @scaler.run_loop
     end
   end
