+++ spec/lib/heroku_dyno_manager_spec.rb
@@ -24,22 +24,4 @@ describe Dynosaur::HerokuDynoManager do
 
   end
 
-  it "should re-check before modifying" do
-    manager = Dynosaur::HerokuDynoManager.new(@api_key, @app_name, true, 30)
-    manager.set(2)
-
-    current = manager.get_current_value
-    current.should eql 2
-    manager.retrievals.should eql 1
-
-    manager.ensure_value(2)
-    current = manager.get_current_value
-    current.should eql 2
-    manager.retrievals.should eql 2
-
-    manager.ensure_value(3)
-    current = manager.get_current_value
-    current.should eql 3
-    manager.retrievals.should eql 3
-    end
 end
